Server/requirements.txt
fastapi
uvicorn
pytesseract
Pillow
